Use normalized JSON representation internally
Allows the input spec to use either atoms or binaries as map keys.
E.g., if parsing an external JSON file instead of using inline Erlang
maps in the handler modules, the default behaviour would be not to
create atoms for keys. Similarly, all string values are normalized to
binaries. This makes it easy to look up a key and check its value.